Revert "Merge branch 'v6.0' of github.com:10gen/mongo into v6.0"
This reverts commit e24af2270e6bc1d435845c8fdd02a1eb24155da2, reversing
changes made to ecf315a92efb6283a4c8f3fd079dd81398563fa6.

+++ src/mongo/db/pipeline/pipeline_d.cpp
@@ -1451,9 +1451,7 @@ PipelineD::buildInnerQueryExecutorGeneric(const MultipleCollectionAccessor& coll
                                     // This produces {$const: maxBucketSpanSeconds}
                                     make_intrusive<ExpressionConstant>(
                                         expCtx.get(),
-                                        Value{static_cast<long long>(
-                                                  unpack->getBucketMaxSpanSeconds()) *
-                                              1000}))),
+                                        Value{unpack->getBucketMaxSpanSeconds() * 1000}))),
                             expCtx);
                         pipeline->_sources.insert(
                             unpackIter,
